+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## DEV
+## v4.3.0 - 2020-08-12
 
 - General Commander's minions are now dealt as cards to your inventory